Undo pep-8 autoformated changes to geometric_test module

+++ pyrr/geometric_tests.py
@@ -16,7 +16,6 @@ TODO: line_intersect_line
 TODO: line_segment_intersect_line_segment
 """
 
-
 @all_parameters_as_numpy_arrays
 def point_intersect_line(point, line):
     """Calculates the intersection point of a point and aline.
@@ -33,7 +32,6 @@ def point_intersect_line(point, line):
     if np.count_nonzero(cross) > 0:
         return None
     return point
-
 
 @all_parameters_as_numpy_arrays
 def point_intersect_line_segment(point, line):
@@ -58,7 +56,6 @@ def point_intersect_line_segment(point, line):
     if dot < 0.0 or dot > squared_length:
         return None
     return point
-
 
 @all_parameters_as_numpy_arrays
 def point_intersect_rectangle(point, rect):
@@ -122,7 +119,6 @@ def ray_intersect_plane(ray, plane, front_only=False):
     t = (pd - p0_n) / rd_n
     return ray[0] + (ray[1] * t)
 
-
 @all_parameters_as_numpy_arrays
 def point_closest_point_on_ray(point, ray):
     """Calculates the point on a ray that is closest to a point.
@@ -145,7 +141,6 @@ def point_closest_point_on_ray(point, ray):
     relative_point = (point - ray[0])
     t = vector.dot(relative_point, normalized_n)
     return ray[0] + (normalized_n * t)
-
 
 @all_parameters_as_numpy_arrays
 def point_closest_point_on_line(point, line):
@@ -173,7 +168,6 @@ def point_closest_point_on_line(point, line):
     rl = vector.normalize(rl)
     dot = vector.dot(rl, rp)
     return line[0] + (rl * dot)
-
 
 @all_parameters_as_numpy_arrays
 def point_closest_point_on_line_segment(point, segment):
@@ -208,7 +202,6 @@ def point_closest_point_on_line_segment(point, segment):
     # perform the same calculation as closest_point_on_line
     return segment[0] + (rl * dot)
 
-
 @all_parameters_as_numpy_arrays
 def vector_parallel_vector(v1, v2):
     """Checks if two vectors are parallel.
@@ -222,7 +215,6 @@ def vector_parallel_vector(v1, v2):
     cross = vector3.cross(v1, v2)
     return 0 == np.count_nonzero(cross)
 
-
 @all_parameters_as_numpy_arrays
 def ray_parallel_ray(ray1, ray2):
     """Checks if two rays are parallel.
@@ -233,8 +225,7 @@ def ray_parallel_ray(ray1, ray2):
     """
     # we use a cross product in-case the ray direction
     # isn't unit length
-    return vector_parallel_vector(ray1[1], ray2[1])
-
+    return vector_parallel_vector(ray1[ 1 ], ray2[ 1 ])
 
 @all_parameters_as_numpy_arrays
 def ray_coincident_ray(ray1, ray2):
@@ -265,7 +256,6 @@ def ray_coincident_ray(ray1, ray2):
         return True
     return False
 
-
 @all_parameters_as_numpy_arrays
 def ray_intersect_aabb(ray, aabb):
     """Calculates the intersection point of a ray and an AABB
@@ -284,16 +274,17 @@ def ray_intersect_aabb(ray, aabb):
     # so where the ray direction value is 0.0, just use infinity
     # which is what we want anyway
     direction = ray[1]
-    dir_fraction = np.empty(3, dtype=ray.dtype)
+    dir_fraction = np.empty(3, dtype = ray.dtype)
     dir_fraction[direction == 0.0] = np.inf
     dir_fraction[direction != 0.0] = np.divide(1.0, direction[direction != 0.0])
 
-    t1 = (aabb[0, 0] - ray[0, 0]) * dir_fraction[0]
-    t2 = (aabb[1, 0] - ray[0, 0]) * dir_fraction[0]
-    t3 = (aabb[0, 1] - ray[0, 1]) * dir_fraction[1]
-    t4 = (aabb[1, 1] - ray[0, 1]) * dir_fraction[1]
-    t5 = (aabb[0, 2] - ray[0, 2]) * dir_fraction[2]
-    t6 = (aabb[1, 2] - ray[0, 2]) * dir_fraction[2]
+    t1 = (aabb[0,0] - ray[0,0]) * dir_fraction[ 0 ]
+    t2 = (aabb[1,0] - ray[0,0]) * dir_fraction[ 0 ]
+    t3 = (aabb[0,1] - ray[0,1]) * dir_fraction[ 1 ]
+    t4 = (aabb[1,1] - ray[0,1]) * dir_fraction[ 1 ]
+    t5 = (aabb[0,2] - ray[0,2]) * dir_fraction[ 2 ]
+    t6 = (aabb[1,2] - ray[0,2]) * dir_fraction[ 2 ]
+
 
     tmin = max(min(t1, t2), min(t3, t4), min(t5, t6))
     tmax = min(max(t1, t2), max(t3, t4), max(t5, t6))
@@ -314,7 +305,6 @@ def ray_intersect_aabb(ray, aabb):
     point = ray[0] + (ray[1] * t)
     return point
 
-
 @all_parameters_as_numpy_arrays
 def point_height_above_plane(point, plane):
     """Calculates how high a point is above a plane.
@@ -334,7 +324,6 @@ def point_height_above_plane(point, plane):
     Just perform Pn . [XYZ1]
     """
     return np.dot(plane, [point[0], point[1], point[2], 1.0])
-
 
 @all_parameters_as_numpy_arrays
 def point_closest_point_on_plane(point, plane):
@@ -359,7 +348,6 @@ def point_closest_point_on_plane(point, plane):
     d = np.dot(p, n)
     qn = np.dot(point, n)
     return point + (n * (d - qn))
-
 
 @all_parameters_as_numpy_arrays
 def sphere_does_intersect_sphere(s1, s2):
@@ -387,7 +375,6 @@ def sphere_does_intersect_sphere(s1, s2):
         return False
     return True
 
-
 @all_parameters_as_numpy_arrays
 def sphere_penetration_sphere(s1, s2):
     """Calculates the distance two spheres have penetrated
@@ -412,7 +399,6 @@ def sphere_penetration_sphere(s1, s2):
     if penetration <= 0.0:
         return 0.0
     return penetration
-
 
 @all_parameters_as_numpy_arrays
 def ray_intersect_sphere(ray, sphere):
